% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercorr_cat_nb.R
\name{intercorr_cat_nb}
\alias{intercorr_cat_nb}
\title{Calculate Intermediate MVN Correlation for Ordinal - Negative Binomial Variables: Correlation Method 1}
\usage{
intercorr_cat_nb(rho_cat_nb = NULL, marginal = list(), size = NULL,
  mu = NULL, p_zinb = 0, nrand = 1e+05, seed = 1234)
}
\arguments{
\item{rho_cat_nb}{a \code{k_cat x k_nb} matrix of target correlations among ordinal and Negative Binomial variables; the NB variables
should be ordered 1st regular, 2nd zero-inflated}

\item{marginal}{a list of length equal to \code{k_cat}; the i-th element is a vector of the cumulative
probabilities defining the marginal distribution of the i-th variable;
if the variable can take r values, the vector will contain r - 1 probabilities (the r-th is assumed to be 1)}

\item{size}{a vector of size parameters for the Negative Binomial variables (see \code{\link[stats]{dnbinom}}); the order should be
1st regular NB variables, 2nd zero-inflated NB variables}

\item{mu}{a vector of mean parameters for the NB variables; order the same as in \code{size}}

\item{p_zinb}{a vector of probabilities of structural zeros (not including zeros from the NB distribution) for the zero-inflated NB variables
(see \code{\link[VGAM]{dzinegbin}}); if \code{p_zinb} = 0, \eqn{Y_{nb}} has a regular NB distribution;
if \code{p_zinb} is in \code{(-prob^size/(1 - prob^size),} \code{0)}, \eqn{Y_{nb}} has a zero-deflated NB distribution and \code{p_zinb}
is not a probability; if \code{p_zinb = -prob^size/(1 - prob^size)}, \eqn{Y_{nb}} has a positive-NB distribution (see
\code{\link[VGAM]{dposnegbin}}); if \code{length(p_zinb) < length(size)}, the missing values are set to 0 (and ordered 1st)}

\item{nrand}{the number of random numbers to generate in calculating the bound (default = 10000)}

\item{seed}{the seed used in random number generation (default = 1234)}
}
\value{
a \code{k_cat x k_nb} matrix whose rows represent the \code{k_cat} ordinal variables and columns represent the
    \code{k_nb} Negative Binomial variables
}
\description{
This function calculates the \code{k_cat x k_nb} intermediate matrix of correlations for the \code{k_cat} ordinal (\eqn{r >=
    2} categories) and \code{k_nb} Negative Binomial variables required to produce the target correlations in \code{rho_cat_nb}. It extends the method of Amatya & Demirtas (2015, \doi{10.1080/00949655.2014.953534})
    to ordinal - Negative Binomial pairs and allows for regular or zero-inflated NB variables.  Here, the intermediate correlation between Z1 and Z2 (where Z1 is the standard normal variable
    discretized to produce an ordinal variable Y1, and Z2 is the standard normal variable used to generate a Negative Binomial
    variable via the inverse CDF method) is calculated by dividing the target correlation by a correction factor.  The
    correction factor is the product of the upper Frechet-Hoeffding bound on the correlation between a Negative Binomial variable
    and the normal variable used to generate it and a simulated GSC upper bound on the correlation between an ordinal variable and the normal variable used to generate it (see Demirtas & Hedeker, 2011,
    \doi{10.1198/tast.2011.10090}).  The function is used in \code{\link[SimCorrMix]{intercorr}} and \code{\link[SimCorrMix]{corrvar}}.
    This function would not ordinarily be called by the user.
}
\references{
Please see references for \code{\link[SimCorrMix]{intercorr_cat_pois}}
}
\seealso{
\code{\link[SimCorrMix]{intercorr}}, \code{\link[SimCorrMix]{corrvar}}
}
\keyword{NegativeBinomial}
\keyword{correlation}
\keyword{method1}
\keyword{ordinal}
