% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercorr_cat_pois.R
\name{intercorr_cat_pois}
\alias{intercorr_cat_pois}
\title{Calculate Intermediate MVN Correlation for Ordinal - Poisson Variables: Correlation Method 1}
\usage{
intercorr_cat_pois(rho_cat_pois = NULL, marginal = list(), lam = NULL,
  p_zip = 0, nrand = 1e+05, seed = 1234)
}
\arguments{
\item{rho_cat_pois}{a \code{k_cat x k_pois} matrix of target correlations among ordinal and Poisson variables; the Poisson variables
should be ordered 1st regular, 2nd zero-inflated}

\item{marginal}{a list of length equal to \code{k_cat}; the i-th element is a vector of the cumulative
probabilities defining the marginal distribution of the i-th variable;
if the variable can take r values, the vector will contain r - 1 probabilities (the r-th is assumed to be 1)}

\item{lam}{a vector of lambda (mean > 0) constants for the regular and zero-inflated Poisson variables (see \code{stats::dpois});
the order should be 1st regular Poisson variables, 2nd zero-inflated Poisson variables}

\item{p_zip}{a vector of probabilities of structural zeros (not including zeros from the Poisson distribution) for the
zero-inflated Poisson variables (see \code{VGAM::dzipois}); if \code{p_zip} = 0, \eqn{Y_{pois}} has a regular Poisson
distribution; if \code{p_zip} is in (0, 1), \eqn{Y_{pois}} has a zero-inflated Poisson distribution;
if \code{p_zip} is in \code{(-(exp(lam) - 1)^(-1), 0)}, \eqn{Y_{pois}} has a zero-deflated Poisson distribution and \code{p_zip}
is not a probability; if \code{p_zip = -(exp(lam) - 1)^(-1)}, \eqn{Y_{pois}} has a positive-Poisson distribution
(see \code{VGAM::dpospois}); if \code{length(p_zip) < length(lam)}, the missing values are set to 0 (and ordered 1st)}

\item{nrand}{the number of random numbers to generate in calculating the bound (default = 10000)}

\item{seed}{the seed used in random number generation (default = 1234)}
}
\value{
a \code{k_cat x k_pois} matrix whose rows represent the \code{k_cat} ordinal variables and columns represent the \code{k_pois} Poisson variables
}
\description{
This function calculates a \code{k_cat x k_pois} intermediate matrix of correlations for the \code{k_cat} ordinal (\eqn{r >=
    2} categories) and \code{k_pois} Poisson variables required to produce the target correlations in \code{rho_cat_pois}. It extends the method of Amatya & Demirtas (2015, \doi{10.1080/00949655.2014.953534})
    to ordinal - Poisson pairs and allows for regular or zero-inflated Poisson variables.
    Here, the intermediate correlation between Z1 and Z2 (where Z1 is the standard normal variable discretized to produce an
    ordinal variable Y1, and Z2 is the standard normal variable used to generate a Poisson variable via the inverse CDF method) is
    calculated by dividing the target correlation by a correction factor.  The correction factor is the product of the
    upper Frechet-Hoeffding bound on the correlation between a Poisson variable and the normal variable used to generate it
    and a simulated GSC upper bound on the correlation between an ordinal variable and the normal variable used to generate it (see
    Demirtas & Hedeker, 2011, \doi{10.1198/tast.2011.10090}).  The function is used in \code{\link[SimCorrMix]{intercorr}} and
    \code{\link[SimCorrMix]{corrvar}}.  This function would not ordinarily be called by the user.
}
\references{
Amatya A & Demirtas H (2015). Simultaneous generation of multivariate mixed data with Poisson and normal marginals.
    Journal of Statistical Computation and Simulation, 85(15):3129-39. \doi{10.1080/00949655.2014.953534}.

Demirtas H & Hedeker D (2011). A practical way for computing approximate lower and upper correlation bounds.
    American Statistician, 65(2):104-109. \doi{10.1198/tast.2011.10090}.

Frechet M (1951). Sur les tableaux de correlation dont les marges sont donnees.  Ann. l'Univ. Lyon SectA, 14:53-77.

Hoeffding W. Scale-invariant correlation theory. In: Fisher NI, Sen PK, editors. The collected works of Wassily Hoeffding.
    New York: Springer-Verlag; 1994. p. 57-107.

Yahav I & Shmueli G (2012). On Generating Multivariate Poisson Data in Management Science Applications. Applied Stochastic
    Models in Business and Industry, 28(1):91-102. \doi{10.1002/asmb.901}.

Yee TW (2017). VGAM: Vector Generalized Linear and Additive Models. \cr \url{https://CRAN.R-project.org/package=VGAM}.
}
\seealso{
\code{\link[SimCorrMix]{intercorr}}, \code{\link[SimCorrMix]{corrvar}}
}
\keyword{Poisson}
\keyword{correlation}
\keyword{method1}
\keyword{ordinal}
